M2-4835: Replace "prettier" with "plugin:prettier/recommended"
This plugin adds prettier as an ESLint rule, so that files not properly formatted according to the prettier config will get flagged as ESLint errors.

I also moved the configuration to the end of the array, to allow it to turn off preceding rules (as it does).

Finally, I removed the remaining rules that conflict with those from `eslint-config-prettier`, namely "quotes", "no-trailing-spaces", and "arrow-parens". I also moved "jsx-quotes" to the prettier config