M2-4835: Linting rules for `mindlogger-admin` (#1472)
This PR creates a GitHub Action to run linting tasks on pull requests submitted to the repository. As part of the ticket, I took the opportunity to update the ESlint rules on the project. These updates have raised a number of linting issues, but **won't block** the merging of pull requests. While I've addressed some of the smaller items (and any error ones), I expect that these will be addressed in the normal course of feature development so that we can eventually require linting checks to pass in order to merge PRs.

Here's a summary of what changed:

## Conflicting ESLint configurations

There are currently two ESLint configurations in the project:
- An `eslintConfig` property in the `package.json` file
- A `.eslintrc` file

[ESLint documentation](https://eslint.org/docs/latest/use/configure/configuration-files) says that only one of these is read, and the `.eslintrc` takes precedence. Thus, I've removed the ignored configuration since it wasn't being used.

## Prettier Code Formatting

Prettier has been configured in the project as the code formatter of choice, but it is currently conflicting with a number of code formatting ESLint rules. The prettier ESLint plugin disables these rules if it is placed as an inherited configuration at the end of the `extends` array, so I have made this change. The prettier ESLint plugin also enables Prettier as an ESLint rule, so that code formatting issues become errors. This should be okay, since Prettier fixes them for you.

In addition, I've disabled/removed some other code formatting related rules that aren't explicitly handled by the Prettier plugin. These include:
- `padding-line-between-statements`
- `no-extra-semi`
- `jsx-quotes` - I've moved this one to the Prettier config with `jsxSingleQuote`

## Opinionated changes

There are a number of other changes I've made that I think will allow for improved code quality. These are:
- Turning on the `react-hooks/exhaustive-deps` rule
- Adding the `unused-imports` ESLint plugin
- Turning on the `@typescript-eslint/no-non-null-assertion` rule