M2-5443: populate Applet > Activities screen
Implementation is based on Table framework (using `formatRow`) in
anticipation of future table view option for this screen (not due for
delivery until later).

For now, activity properties that are not yet available from API – i.e.,
participant count, compliance, trend, last completion data – are stubbed
as `null`. Included TODOs linking to JIRA tickets for filtering/sorting
and activity actions.

Add new SVG sprites and update `export` sprite, which is only slightly
tweaked to stay pixel-perfect with Figma. Confirmed that existing uses
of this sprite in the app continue to render as expected.