M2-5921: Fix release event type (#1626)
This PR follows up https://github.com/ChildMindInstitute/mindlogger-admin/pull/1617 by changing the release event type to `published` instead of `created`. The `created` event type doesn't capture releases that were previously saved as draft, so some releases would be missed.

This change also brings the workflow file to the develop branch. This is important because that's where the tagging takes place and the workflow file won't be triggered if it doesn't exist on that branch.