Update the list of available team members
Only owners and managers are eligible for inclusion into the list of team members in the various dropdown lists. This change makes that change.

I also modify the wording of the warning message to remove reference to limited accounts. This is because it also applies to full account participants in phase 1.

Finally, I fully disable the checkbox since its state is entirely controlled by the type of user in the dropdown fields