feat(mi): Updated take now modal form (M2-6480) (#1728)
This PR updates the form on the TakeNowModal to match the new designs in Figma. The form now has a new checkbox, a new dropdown field, and the order of the fields have been rearranged. The tooltips have also been removed.

Submission of the form will now send these query parameters to the respondent web app:
- `startActivityOrFlow` - The ID of the activity
- `sourceSubjectId` - The subject ID of the user in the "**Who will be providing the responses?**" dropdown
- `targetSubjectId` - The subject ID of the user in the "**Who are the responses about?**" dropdown
- `subjectId` - The same as `targetSubjectId` for backwards compatibility with the respondent web app
- `respondentId` - The user ID of the user in the "**Who will be providing the responses?**" dropdown when the checkbox is checked, otherwise the user ID of the user in the "**Who will be inputting the responses?**" dropdown