fix: add keyExtractor support to tables
It's bad practice to striclty use array index for React key props, which
is what Table and DashboardTable were using, causing residual styling
to be applied to the wrong element when items shift. Added support for
`keyExtractor`s to both Table and Dashboard table so that row IDs can be
used as keys instead. Required adding support for hidden (no content)
columns.