fix: fix typing/thumbnail for Take Now w/ flows
Updated `TakeNowModal` to expect either an Activity or an ActivityFlow
rather than implicitly using type coersion to satisfy TS. Also fix the
thumbnail shown in the modal to reflect the multiple activities included
in the selected flow.

These changes are required to properly report to analytics whether an
activity or activity flow has been requested for Take Now.