feat: update `useLogout` to store soft-lock data
Also update `TakeNowModal` to take advantage of the hook's new prop.

Note that because the signature of the function returned by `useLogout`
has changed, we had to wrap it in a closure in `LockForm` so as to keep
TypeScript happy.

Also note that `LockForm` is a legacy component that is not currently in
use, and may not be used in the future depending on whether this
implementation of soft-lock proves successful as an alternative
approach. (`LockForm` was created to be activated after 15 minutes of
inactivity, but never fully implemented.)