feat: add `ActiviesList` & `ActivityCheckbox`
`ActivitesList` renders the scrollable list of avaialble activities and
activity flows, allowing the user to click rows to select/deselect
items. Required adding support for "size" prop to `Chip`s.

`ActivityCheckbox` is used in a couple places during activity selection,
so it made sense to create a reusable component. Required a new SVG
sprite, as well as applying `currentColor` to existing sprite fill
during use instead of in definition, due to the way SVG fill is
overridden using CSS in `Button` components.