feat: add drawer pane infra & use `ActivitiesList`
Add lots of infra to `ActivityAssignDrawer`:
- form state management
- step management (selection vs. review) and transitions between steps
- footer button transitions
- populating `ActivitiesList` from the API
- selecting/deselecting activities

A note about how to maintain fade transitions between steps is added as
a comment. This approach is probably the simplest way to achieve fades,
but relies on one of the panes having a guaranteed taller content height
than the other. If this approach is deemed too fragile, that can be
addressed when M2-7261 is implemented.

Added some global theme styles that seemed appropriate. Worth noting is
the introduction of the `outlined` style variant for `IconButton`s,
which has been in the designs since ML1 but never implemented. Since
`IconButton`s lack a `variant` prop, used the `color` prop and a
custom-named color variant.