refactor: add `isTeamMember` prop to options
Use new prop `isTeamMember` to identify when a dropdown option is part
of a team. This prop is initialized on creation based on the same logic
as before – checking whether the tag is "Team". This will enable us to
render the "Self" dropdown option without a "Team" tag beside it
when that "self" subject is a team member (as per designs), but still
retain its position in the "Team" grouping of dropdowns.