fix: consolidate Team Member snippet render logic
When passing a Team Member to a `ParticipantSnippet`, there was logic
that only existed in `ParticipantDropdown` to render the right fields in
the right order. Moved that logic into the snippet component so that
Team Members are always rendered in the same way wherever the snippet is
used.

Also rename `ParticipantDropdown` i18n keys to account for its
independence from Take Now.