fix: update dropdown populating logic to follow AC
After re-reading the task's AC, discovered that the right dropdowns
should exclude Team Members, so updated that. Also, the Self option now
appears in its own unlabelled group at the top, regardless of where it
may have occurred naturally in the results, to follow designs.

Secondly, only the "Add Respondent" error state is present in the
designs (no "Add Subject", so removed that), and it's only shown if the
selected subject is a limited account. There's also more nuance around
whether to populate the right column with "self" based on left column
selection.