feat: MI Assign: Specify Activities & Respondents (M2-7135, M2-7507) (#1870)
## feat: Add ActivityAssignDrawer

## feat: add HelpPopup

Also create reusable styled component `BorderedOverflow` that renders a
vertically scrollable container that displays a light grey top/bottom
border whenever overflowing content exists beyond that edge.

Fix font weight of default submit button of modals; it is never
indicated in Figma anywhere to override default button variant styling
for modals. Probably a relic of outdated styling.

## feat: add `ActivitiesList` & `ActivityCheckbox`

`ActivitiesList` renders the scrollable list of available activities and
activity flows, allowing the user to click rows to select/deselect
items. Required adding support for "size" prop to `Chip`s.

`ActivityCheckbox` is used in a couple places during activity selection,
so it made sense to create a reusable component. Required a new SVG
sprite, as well as applying `currentColor` to existing sprite fill
during use instead of in definition, due to the way SVG fill is
overridden using CSS in `Button` components.

## refactor: add `hydrateActivityFlows` utility

There were already 2 instances where this is happening in the Dashboard,
and with `ActivityAssignDrawer` we will have a 3rd, so it made sense to
make a shared utility. This utility populates Activity Flows with their
activity entities (rather than just `activityIds`), which is needed to
render their thumbnails.

## feat: add drawer pane infra & use `ActivitiesList`

Add lots of infra to `ActivityAssignDrawer`:
- form state management
- step management (selection vs. review) and transitions between steps
- footer button transitions
- populating `ActivitiesList` from the API
- selecting/deselecting activities

A note about how to maintain fade transitions between steps is added as
a comment. This approach is probably the simplest way to achieve fades,
but relies on one of the panes having a guaranteed taller content height
than the other. If this approach is deemed too fragile, that can be
addressed when M2-7261 is implemented.

Add some global theme styles that seemed appropriate. Worth noting is
the introduction of the `outlined` style variant for `IconButton`s,
which has been in the designs since ML1 but never implemented. Since
`IconButton`s lack a `variant` prop, used the `color` prop and a
custom-named color variant.

## feat: add banner support to Activity Assign

Adapt existing global Banners infrastructure to `ActivityAssignDrawer`,
introducing a hook to make it convenient to manage display of local
banners without overcomplicating the component.

Also fix a layout bug in the Banner component (they were never being
rendered properly when there was no close button). Add support to
customize the icon as well.

Also fix a styling/layout issue with the footer reveal.

## refactor: LabeledUserDropdown ⇒ ParticipantDropdown

Refactoring this as a shared component, including:

- Rename it to something that better describes its purpose
- Move functions, utilities, and types specific to `ParticipantDropdown`s
into own hook, which can be used more portably rather than strictly in
the Take Now modal.
- Extract more dropdown content into `TakeNowModal` component by
creating `TakeNowDropdown`, a Take Now-specific wrapper for
`ParticipantDropdown`

## refactor: add `isTeamMember` prop to options

Use new prop `isTeamMember` to identify when a dropdown option is part
of a team. This prop is initialized on creation based on the same logic
as before – checking whether the tag is "Team". This will enable us to
render the "Self" dropdown option without a "Team" tag beside it
when that "self" subject is a team member (as per designs), but still
retain its position in the "Team" grouping of dropdowns.

## feat: add `AssignmentsTable` & validation

Add `AssignmentsTable` component to Activity Assign, which lists activity
assignments using `ParticipantDropdown` (wrapped by
`AssignmentDropdown`). Changes include:

- display # of self-report & multi-informant assignments
- update form completion state that takes into account assignment rows
- `onSubmit` validation, which now checks for duplicate assignment rows
- add ability to style invalid rows in error state, by adding a new
  special prop to the `Row` type of both `DashboardTable` and `Table`
- optional "empty value" state for `ParticipantDropdown`, when a value
  is unselected and the user needs to be alerted
- replace matching `ParticipantDropdownOption` in target subject
  dropdown with "Self" when a respondent has been selected, and ensure
  `ParticipantDropdown` reacts to changes in the `options` prop to
  reflect this change
- update code comments for improved clarity in `ActivityAssignDrawer`

## fix: improve UX of `ParticipantDropdown`

Ran into various UX issues in `ParticipantDropdown` when doing a lot
of rapid searching, focusing, and blurring of the component. Attempted
to fix most of these using some more nuanced focus state management.
Felt more natural to clear the dropdown when it loses focus so enabled
that by default. Tested to confirm these changes were appropriate for
Take Now modal as well.

## fix: improve incomplete assignments validation

Pending some final recommendations from design, updated assignments
validation to allow the user to proceed to Review step if there are
empty assignment rows. This allows the user to clear duplicate
assignments, without a "delete row" button (currently not present in
designs). Empty rows are deleted when validation succeeds, as per
suggestion made in [this Figma thread](https://www.figma.com/design/6BNx63peNVzrHAF1FoZ78x?node-id=9404-158430#881972698).

## style: `ParticipantDropdown` design alignment

Display `ParticipantSnippet` when the component is unfocused. Fix
size of dropdown & clear icons, and padding of the text field and
snippet to accommodate.

Add new "full" variant to ParticipantDropdown for varied styling in
AssignmentsTable. Required adding hover effect and a number of spacing
adjustments.

## style: add Limited Account icon to dropdowns

## test: fix broken tests

Participants screen, which now includes ActivityAssignDrawer, needs
some extra HTTP requests handled for their tests to continue passing.

## test: add `ActivitiesList` tests

## test: add tests for `AssignmentsTable`

In adding tests, fix some other issues:
- add aria-attribute to invalid table rows for a11y
- make correction to logic of conditionally issued API requests during
  search; issues with it were revealed during simultaneous test runs
- eliminate redundant dropdown search requests when value is changed
  programmatically
- fix styling

## test: add tests for `HelpPopup`

## test: add tests for `ActivityAssignDrawer`

## fix: ParticipantSnippet consistency (M2-7507)

Update Applet Overview to align with how `ParticipantSnippet`s are
rendered elsewhere in the app. This small change was included in this PR
because it relates to making `ParticipantSnippet` usage consistent,
which was an essential part of the work in this PR.

## test: add `ActivityAssignDrawer` submission tests

Also separate shared testing utilities into a `.test-utils.ts` file.