feat: list only assigned activities on PDP
Update the `Participant/Activities` screen to only show activities for
which the participant has assignments either as the respondent or the
target subject.

Update tests to reflect this change. Also move definition of
`mockedOwnerSubject` to the shared `mock.ts` file, and align the
subject details in the `preloadedState` with that subject. This required
making small tweaks to two tests.