feat: add `ActivityUnassignDrawer`
This component includes both the unassign panel (when there are multiple
assignments), and the unassign confirmation popup (when there is only
one assignment, or after submitting the form in the panel).

Enhanced `useAsync` to be able to read the arguments passed for the
current async call after it completes, to be able to populate the
success banner when the required info is not returned by the response.

Unit tests still pending for `ActivityUnassignDrawer` component.