feat: incorporate `ActivityUnassignDrawer` ⇒ PDP
In making this change, we still need to fetch the
`/activities/applet/{applet_id}` endpoint to populate the Unassigned
section, per the original PDP designs. This is obviously costly in
performance, but this version of the PDP is only here temporarily to be
able to properly test the unassign flow.