feat: Unassign Activities (M2-7277, M2-7547) (#1924)
* feat: add "Unassign Activity" menu item

Add "Unassign Activity" menu item with visibility and disabled state
reflecting assignment permission and status. Adjust other menu items
accordingly.

When clicked, "Unassign Activity" currently only sets a state variable –
the unassign drawer/popup itself is coming in a subsequent commit.

Required introducing API activity/flow types that include `assignments`
property to determine whether participant is assigned to each activity/
flow. Also opted to strongly type the `/activities/applet/{applet_id}`
endpoint response.

* refactor: share assign drawer subcomponent & hook

Move the `AssignmentCounts` subcomponent of `ActivityAssignDrawer` to a
shared location to be eventually consumed by `ActivityUnassignDrawer`.
Also make its banners hook generic, move it to the shared hooks folder,
and rename it to `useModalBanners`.

* feat: add `ActivityUnassignDrawer`

This component includes both the unassign panel (when there are multiple
assignments), and the unassign confirmation popup (when there is only
one assignment, or after submitting the form in the panel).

Enhanced `useAsync` to be able to read the arguments passed for the
current async call after it completes, to be able to populate the
success banner when the required info is not returned by the response.

* feat: incorporate `ActivityUnassignDrawer` ⇒ PDP

In making this change, we still need to fetch the
`/activities/applet/{applet_id}` endpoint to populate the Unassigned
section, per the original PDP designs. This is obviously costly in
performance, but this version of the PDP is only here temporarily to be
able to properly test the unassign flow.

* fix: remove useless Search field in PDP header

* fix: reduce `:first-child` console warnings

* test: add tests for `ActivityUnassignDrawer`

Also make use of reusable mocks in some existing tests.

* fix: address activity/flow language nuances

Also create shared `FlowChip` component that is reused wherever a flow
is shown in the Dashboard (was also missing from the Take Now modal).