feat: add `ExpandedView` to By Participant tab
In this commit we add logic to the By Participant tab to harness the
new support for expanded views in `ActivityListItem`, and pass an
instance of the `ExpandedView` component to each of those list items.
Data retrieval and refetching is managed in the `ByParticipant`
component, the same place where other API calls are already happening.

Each expanded view's data is stored as a property of a shared state
variable – a map of activities/flows ⇒ subject data. Another shared
state variable maps activities/flows ⇒ loading status, which is used
to display a spinner on the expand/collapse button for each
activity/flow.