fix: move encryption cleanup
Duplicated applets with report server configuration was incorrectly sending an empty password to the Report Server during replication. This was because the encryptionDataRef was getting reset in the finally clause, which was taking place before the report server configuration setup had a chance to complete.

This required a change to move this out of the finally clause, and appropriately handle the cleanup on both success and error callbacks.