fix: Report Score Average Range Calculation (M2-7738)
This PR updates the way the report score score range is calculated and displayed for scoring type "Score". When a subscale is linked that uses _Average_ calculation type, the score range now properly excludes the min and max lookup table scores from the division.

The average score range is determined by finding the min and max average scores from each item, and then finding the min/max between those and the min/max scores in the lookup table respectively.