fix: Cascade system item changes to report scores (M2-8146) (#1965)
This PR creates logic that cascades changes to activity system items between the subscales configuration and the report configuration.

Whenever a subscale with a lookup table is added to an activity, two non-editable system items are added to the activity (`gender_screen` and `age_screen`). These items then become available for printing in reports. Whenever the subscale lookup table is removed/replaced, these items also get removed and re-added to the activity. Previously if these items were selected for printing, and then subsequently removed/replaced, report items would retain a reference to the older versions of these system items and subsequently throw an error when it fails to find them.