fix: assign logic for limited respondents
Ensured the `isLimitedRespondent` is only true if a respondent subject
is even present. This ensures assign/unassign menu items follow correct
logic on the About Participant tab.

Also, for the By Participant tab, checks first if the respondent is a
limited account before hiding the Assign menu item and showing the
Unassign menu item, so that when an activity or flow is set to
auto-assign, the user isn't misled into thinking that auto-assigned
activities/flows could have been assigned to the limited account as the
respondent.