fix: refactor RenderedField to correct implementation errors
`PhrasalTemplateRenderField` had changes made to it by a former engineer
who made some poor choices and did not gracefully welcome feedback (he
is no longer with us), leaving the sole reviewer of his work (myself) to
begrudgingly approve his adequate implementation in the interest of
getting the feature to a point of usability. I took this opportunity to
fix some of his work since I had to edit this file anyway, and I was
already seeing the issues with his choices made evident in the form of
console warnings.

Specifically, state was being set in the render thread, which React will
eventually complain about if it detects that happening. To fix this, I
moved the state management out of the `getRenderedValue` function and
into an effect. I also simplified the state variables representing the
dropdown values, and initialized the values of the `displayMode` and
`itemIndex` properties more granularly and carefully, only when needed.

In addition, updated an effect to set the new `_indexedItemLabel`
internal property to the slider name to be consumed by the preview
popup.

To align with the Figma designs, other UI cleanup includes:
- removing the tooltip from the secondary dropdowns
- removing the `startAdornment`s from the secondary dropdowns
- consistently left-aligning the tooltip of the response item dropdown