fix: change drag & drop from `swap` to `move`
For some reason `swap` was being used when an item was drag & dropped to
a new position. Fixed this by replacing with the `move` method.

Also removed some redundant type checking – `index` is already
guaranteed to be a number.

Finally, made use of the `provided.placeholder` element, which is a
requirement of `Droppable` and sometimes produces warnings if absent.