fix: admin no slide number is included in the preview (M2-8206) (#1978)
* fix: align phrase builder type definition with BE & add metadata property

The API is expecting `itemIndex` to be null, not undefined, for
`item_response` field types, so fixed that. Will make further fixes in
`PhrasalTemplateRenderField` in subsequent commit.

Add internal `_indexedItemLabel` field to make the selected slider
name available in the react-hook-form context, to be consumed by
`PreviewPhrasePopup` (next commit).

* fix: display slider label in preview

Update `PreviewPhrasePopup` to use the internal `_indexedItemLabel`
property for inclusion in the placeholder when needed. This property
will be set in `PhrasalTemplateRenderField` in a subsequent commit.

Also add `key` to each return value of `renderField` to prevent console
warnings.

* fix: update `useCustomFormContext` to allow overriding `shouldDirty` flag

This will be needed in order to use `setValue` to set the new metadata
property without dirtying the form.

* fix: refactor RenderedField to correct implementation errors

`PhrasalTemplateRenderField` had changes made to it by a former engineer
who made some poor choices and did not gracefully welcome feedback (he
is no longer with us), leaving the sole reviewer of his work (myself) to
begrudgingly approve his adequate implementation in the interest of
getting the feature to a point of usability. I took this opportunity to
fix some of his work since I had to edit this file anyway, and I was
already seeing the issues with his choices made evident in the form of
console warnings.

Specifically, state was being set in the render thread, which React will
eventually complain about if it detects that happening. To fix this, I
moved the state management out of the `getRenderedValue` function and
into an effect. I also simplified the state variables representing the
dropdown values, and initialized the values of the `displayMode` and
`itemIndex` properties more granularly and carefully, only when needed.

In addition, updated an effect to set the new `_indexedItemLabel`
internal property to the slider name to be consumed by the preview
popup.

To align with the Figma designs, other UI cleanup includes:
- removing the tooltip from the secondary dropdowns
- removing the `startAdornment`s from the secondary dropdowns
- consistently left-aligning the tooltip of the response item dropdown

* fix: clean up remaining console warnings

* fix: deal with possible out-of-range slider index

This change fixes the issue whereby if you have two Slider Rows items,
A and B, with item B containing more sliders than item A, and you have
item B's last slider selected for display in the Phrase Builder, then
change that field's item from item B to item A, the slider selected in
the dropdown will be resetto a valid option for item A.