test: add skipped items tests & fix `calcScores`
In adding unit tests, some kept failing and it helped me discover
a bug in how I was calling `calcScores` from `getSubscales` (a
refactoring relic 🤦🏻‍♂️). While doing so, I opted to better organize
and logic and typing in `calcScores`, which was needlessly obtuse
and cryptic. In turn this may have fixed some further edge-case bugs.