refactor: create full/limited/team accounts ⇒ RTKQ
Migrate mutation calls for creating account invitations and limited
accounts to RTK Query. (Also remove legacy edit/delete subject mutation
functions, which I forgot to include in the previous commit.)

Popups associated with each of these mutations no longer need to inform
the consuming component that a refetch is needed, since that's handled
automatically by RTK Query's cache invalidation system.

It's worth noting that the Applet Overview screen was also responding to
the `shouldRefetch` flag, refreshing the list of submissions in case a
new participant is added. However, submissions are not immediately
affected by newly added participants, so that refetch was also
superfluous.