chore: update tests
With the changes in this PR, a number of tests needed tweaking to pass:

- The `WorkspaceRespondentsResponse` type becoming more restrictive
(including the new data access info) required updating participant mocks
and associated tests.
- The `allRespondents` preloaded state is no longer used and needed to
be cleaned up. In cases where we now are using RTK Query to return that
data, some of the test logic required tweaks to wait for the mock
fetches to complete, since the data is no longer preloaded in Redux.
- `mockAxios` calls no longer always have `{ signal: undefined }`
passed to them, so that needs to be taken into account where those calls
were made by RTK Query.