fix: Error on Dataviz first load
The Dashboard > Applet screen was deleting the Redux-stored applet data
after it was unmounted, causing moments during navigation to subscreens
when the required encryption data was not available to be able to
decrypt answers, resulting in an error on Dataviz when there is
reasonable network latency.

It appears that applet data was being removed to ensure that it's
refreshed in case `appletId` changed. However, the better approach is to
include `appletId` as a dependency on the fetching callback, which is
possible by passing it as an optional dep to the `usePermissions` hook.

Also, the Dataviz component should handle this case better to prevent
displaying a JS error in the React UI.