feat: Data Exports Updates (M2-9151) (#2073)
This PR introduces changes to both the `responses.csv` and `schedule_history.csv` export files by adding an extra column to each file.

#### responses.csv

The `responses.csv` file now has a new column called `event_history_id` (or `activity_schedule_history_id` when the `enableDataExportRenamingfeature` flag is turned on). This represents the value in the `event_history_id` column in the `answers` database table for the given answer.

It is a nullable field, so the column in the export file should be empty whenever the value in the database table is `null`.

#### schedule_history.csv

The `schedule_history.csv` file now has a new column called `schedule_applet_link_timestamp`, which is the UTC timestamp that a particular version of the applet was linked to a particular version of the a schedule event (given by the `applet_version` and `schedule_version` columns respectively in the same file). It is taken from the `created_at` column of the `applet_events` table in the database.