fix: EMA Fast Follows 2 (M2-9106) (M2-9108) (#2081)
This PR fixes two EMA bugs, as described below

#### M2-9106: In the export the new version is linked for the participant not only the individual but also the default schedule

In the specific scenario where the following steps occur on the same day:
- An applet is created (version 1.1.0)
- A participant is added
- The participant is given an individual schedule
- The applet version is increased (version 2.0.0)
- The schedule history is exported for the participant

The default schedule event for the activity in the applet should not be present in the schedule history export file for the participant with the individual schedule on the new applet version. Therefore the schedule history file should contain the following entries for this participant:
- Default schedule event (applet version 1.1.0)
- Individual schedule event (applet version 1.1.0)
- Individual schedule event (applet version 2.0.0)

This is because the default schedule did not apply when the applet version was increased (as the participant already had an individual schedule at that time), and so it should not be present in the schedule history export file.

#### M2-9108: The whole schedule_history associated with the removed participant is also removed

This issue is mostly fixed with https://github.com/ChildMindInstitute/mindlogger-backend-refactor/pull/1848 (and requires it to be tested), but this PR adds a few things:

- Specifies to the API that soft-deleted participants should be included in the workspace respondents endpoint
- Prevents schedule history from being generated past the date when the participant is removed